feat: added indexes

+++ shop/src/main/java/me/shail/dtos/ReviewDto.java
@@ -6,6 +6,6 @@ public record ReviewDto(
         UUID id,
         String title,
         String description,
-        Long rating
+        Integer rating
 ) {
 }
